[#86] Moved camera stuff to different folders, fixed low end gpu crash.
