+++ UnicornRender/video/include/unicorn/video/Graphics.hpp
@@ -146,7 +146,7 @@ public:
     *  @param  window Output window.
     *  @return Pointer to newly created Renderer, nullptr if any error occured.
     */
-    UNICORN_EXPORT Renderer* SpawnRenderer(system::Window* window, Camera* camera);
+    UNICORN_EXPORT Renderer* SpawnRenderer(system::Window* window, Camera& camera);
 
     /** @brief Binds renderer to window. */
     UNICORN_EXPORT void BindWindowRenderer(system::Window* pWindow, Renderer* pRenderer);
